# speedup all binary functions on avx256, speedup binary square on avx512