… lucene/core/src/java20/org/apache/lucene/internal/vectorization/PanamaVectorizationProvider.java