@rmuir 2023-10-14
Here's the diff of just the commit for this change: https://github.com/apache/lucene/pull/12681/commits/3ec9c26d672262762f4213c827699bf735409eeb