@uschindler 2023-10-14
> We have to think about testing. I don't want to rely upon various hardware for correctness. I think there's a way to alter the code so that we can test the correctness of everything (albeit slowly), e.g. exercise 128, 256, and 512 in our tests.

This can be done in the same way like the "testMode" fall, we should just extend it to cover more cases. You could also pass an override for the bit size instead of true/false.

By default it passes the preferred species, in test mode it could pass others.