@rmuir 2023-10-14
> This can be done in the same way like the "testMode" flag, we should just extend it to cover more cases. You could also pass an override for the bit size instead of true/false.
> 
> By default it passes the preferred species, in test mode it could pass others.

A lot of this code relies on this stuff being `static final` constants. For example multiplication by power of 2 on every single vector, but thats just for starts. I don't want to rewrite the entire code for this, just have a plan.

We may have to get more creative here, such as using system property and doing it with "distribution test" passing jvm flags or similar.

I can also use qemu approach which requires no code changes.