@rmuir 2023-10-14
I tried it out, making species `final` instead of `static final`. performance completely falls apart, slower than scalar impl even. it is a non-option... We should keep everything here static final and just test it differently (e.g. more like integration tests that pass different sysprops/jvm flags)