@rmuir 2023-10-14
@uschindler I did the 'fast integer vectors' override differently, and configured the build to randomize the vector size used for testing.

So it still does the same thing it was doing, but now at least it runs with different vector sizes too. Of course they will be slow, but it is for testing code correctness and the test is reasonable.