@rmuir 2023-10-15
for me, when investigating a modification, this works easily enough: 

```console
$ for bits in 128 256 512; do ./gradlew -p lucene/core test --tests TestVectorUtilSupport -Dtests.vectorsize=$bits; done
```