@ChrisHegarty 2023-10-15
> for me, when investigating a modification, this works easily enough:
> 
> ```
> $ for bits in 128 256 512; do ./gradlew -p lucene/core test --tests TestVectorUtilSupport -Dtests.vectorsize=$bits; done
> ```

I love it!  I'm noodling on something in this area too, and having this will make my life soooooo much better.